test: move AMMV panel fixtures to conftest
